Move `pipeline/llm/*` to `llm/*`
